requirements.txt
discord.py
flask
discord.py
flask
yt-dlp
PyNaCl
aiohttp
python-dotenv
